Focus browser hydration fix
